refactor(server): extract auth into src/server/auth.js
First dependency-heavy extraction. Pure move, no behavior change. Relocates the
Clerk JWT + Forge API-key middleware and the shared auth constants:

  - clerkJWKS, CLERK_JWKS_URL, SUPER_ADMIN_IDS
  - verifyBrandAccess, hashApiKey, lookupApiKey, requireAuth,
    requireApiKeyScope, softAuth, mcpAuth

auth.js owns the two shared constants (clerkJWKS, SUPER_ADMIN_IDS) and exports
them, because server.js also reads them directly — 4 inline jwtVerify handlers
use clerkJWKS, ~8 spots check SUPER_ADMIN_IDS. server.js re-imports the
middlewares + those constants; keeps `jwtVerify` from jose for the inline
handlers; drops `createRemoteJWKSet` (only used to build clerkJWKS, now in
auth.js). auth.js imports pool (db.js), jose, and crypto's createHash.

This is exactly the move where a missed re-import is invisible to node --check —
so the no-undef gate is the proof: `npm run lint` is clean, confirming every
moved symbol resolves in both files. Route inventory unchanged (213), 125
requireAuth call sites untouched.

Adds test/auth.test.js (hashApiKey known-vector + determinism;
requireApiKeyScope scope/brand allowlist logic). Full suite: 32 tests green.
